Implement List.cpp Methods

+++ COMP280_Assignment2/COMP280_Assignment2/COMP280_Assignment2/List.h
@@ -23,7 +23,7 @@ public:
     void Delete(ElementType x);
     void Display();
     int Sum();
-    int Average();
+    double Average();
     
 private:
     node* first;
